Computer navigation w Gamepad, Fix bugs from bug report

+++ Assets/Scripts/ScriptableObjects/ProductSO/ProductSO.cs
@@ -6,8 +6,8 @@ using UnityEngine.AddressableAssets;
 public class ProductSO : ScriptableObject {
     [Header("Global variables")]
     public new string name;
-    public float initialPrice;
-    public float price;
+    public int initialPrice;
+    public int price;
     public float cookingTime;
     public int nbCreated;
     public float recipePrice;
